+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New `HeaderConstant`: `PreTrigger`, `PostTrigger`, and `CurrentCalibre`.
 
+### Changed
+
+- Frequency should be given in `TOML` via a `FrequencySetpoint`, even though
+  it can still be given as a `MultipactorTest` argument.
+
 ## [2.0.0] - 2026-06-03
 
 ### Added